+++ internal/db/config.go
@@ -71,9 +71,7 @@ func WithMaxRetries(num int) Option {
 
 func WithNodeIdentity(ident identity.Identity) Option {
 	return func(opts *dbOptions) {
-		if ident != nil {
-			opts.identity = immutable.Some(ident)
-		}
+		opts.identity = immutable.Some(ident)
 	}
 }
